requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
flask-restx==0.5.1
pdfkit==1.0.0
python-docx==0.8.11
html2docx==1.1.0
pydantic==1.8.2
python-multipart==0.0.5
beautifulsoup4==4.10.0
markupsafe==2.0.1
Jinja2==3.0.1
pytest==6.2.5
Flask-Cors==3.0.10
gunicorn==20.1.0
requests==2.28.1